pyproject.toml
[project]
name = "slim-chatui"
description = "Opinionated hard fork of Open WebUI"
authors = [
    { name = "Timothy Jaeryang Baek", email = "tim@openwebui.com" }
]
license = { file = "LICENSE" }
dependencies = [
    "fastapi==0.128.5",
    "uvicorn[standard]==0.40.0",
    "pydantic==2.12.5",
    "python-multipart==0.0.22",

    "python-socketio==5.16.1",
    "cryptography",
    "bcrypt==5.0.0",
    "PyJWT[crypto]==2.11.0",
    "authlib==1.6.7",

    "requests==2.32.5",
    "aiohttp==3.13.2", # do not update to 3.13.3 - broken
    "aiocache",
    "aiofiles",
    "starlette-compress==1.7.0",
    "Brotli==1.1.0",
    "httpx[socks,http2,zstd,cli,brotli]==0.28.1",
    "python-mimeparse==2.0.0",

    "sqlalchemy==2.0.46",
    "alembic==1.18.3",
    "peewee==3.19.0",
    "peewee-migrate==1.14.3",

    "pytz==2025.2",

    "loguru==0.7.3",
    "asgiref==3.11.1",

    "mcp==1.26.0",

    "Markdown==3.10.1",
    "azure-identity==1.25.1",

    "black==26.1.0",
]
readme = "README.md"
requires-python = ">= 3.11, < 3.13.0a1"
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia",
]

[project.optional-dependencies]
postgres = [
    "psycopg2-binary==2.9.11",
]

all = [
    "psycopg2-binary==2.9.11",
]

[project.scripts]
open-webui = "open_webui:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "package.json"
pattern = '"version":\s*"(?P<version>[^"]+)"'

[tool.hatch.build.hooks.custom]  # keep this for reading hooks from `hatch_build.py`

[tool.hatch.build.targets.wheel]
sources = ["backend"]
exclude = [
    ".dockerignore",
    ".gitignore",
    ".webui_secret_key",
    "dev.sh",
    "start.sh",
    "start_windows.bat",
    "webui.db",
    "chroma.sqlite3",
]
force-include = { build = "open_webui/frontend" }

[tool.codespell]
# Ref: https://github.com/codespell-project/codespell#using-a-config-file
skip = '.git*,*.svg,package-lock.json,i18n,*.lock,*.css,*-bundle.js,locales,example-doc.txt,emoji-shortcodes.json'
check-hidden = true
# ignore-regex = ''
ignore-words-list = 'ans'

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]
